app/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
pillow>=11.0.0
torch>=2.1.0
transformers>=4.41.0
huggingface_hub>=0.24.0
safetensors>=0.4.0
